Specify behaviour when mixing APIs
Add wording on whether a user can combine the
explicit graph building API with the record & replay
API on the same modifiable graph object.

It is specified as being allowed for the user to mix
mechanisms, so long as the two mechanisms are used sequentially.
However, it is forbidden if the mechanisms are interleaved and
an exception must be thrown by the implementation.

We decided this because it is not specified in the record & replay
API whether commands are added to the graph eagerly during recording,
or on `queue::end_recording`. When each mechanism is used sequentially
however, the ordering of nodes being added is well defined.

See Issue https://github.com/reble/llvm/issues/11