Expand on profiling semantics
Elaborates on what the semantics are of queries on events
returned by executable graph submission.

The event enters the running state when the first node starts
executing on device, and completes once the last node finishes.

Since a graph could be divided up into multiple PI commands-buffers
that may have work scheduled in-between this could give a pessimistic
view of graph execution time, however I don't think we can guarantee
better.

In the future we could add explicit graph profiling nodes for a
more fine grained view of performance, but that's out of scope
for now.

Closes Issue https://github.com/reble/llvm/issues/52