[SYCL][DOC] Error on simultaneous graph submission
Actions [feedback]
(https://github.com/intel/llvm/pull/5626#discussion_r1150765202)
that we should error when a graph is submitted before a previous
execution has completed, rather than make it conditional on the
backend when we have no query.

Added an issue to the bottom of the spec, since I think for revision
2 we could possibly do better and serialize the graph executions in
the runtime.

Closes https://github.com/reble/llvm/issues/122