[SYCL][Doc] Define interaction with other extensions
Defines the interaction of `sycl_ext_oneapi_graph` with
other extensions that define queue properties or new queue
methods, as these need clarified as to how they relate to
record & replay.

Also ban the use of new handler methods, and queue shortcuts,
in graph nodes. These could be possible to support, but I don't
think the scope of implementing and testing them is feasible for
2024.0 timelines.

Closes Issue https://github.com/reble/llvm/issues/135