[SYCL][Doc] Device query for graphs support
Although we have an emulation mode for when a SYCL backend
doesn't support PI/UR command-buffers, we don't have good implementation coverage
across it's usage on all possible backends.

Providing a support query would allow us to limit the backends we support
while an experimental extension to say CUDA and Level Zero, and maybe
OpenCL. Rather than
having users with FPGA and esimd_emulator trying to use the extension and
crashing because we've not tested these platforms even with emulation
mode.

We could remove this query once the implementation is more mature and
we can trust emulation mode support on all backends, but I don't
think that's the case at the moment.