[SYCL][Graphs] Fix check-sycl target by removing PI command-buffer unit tests
The check-sycl target currently fails to compile our PI unit tests for Level Zero that test PI command-buffers. This exhibits as a fail on Windows CI, but is not a Windows specific issues, and can be recreated on Linux.

The initial set of build fails are due to our test failing to be able to find required headers, this is because Unified Runtime and Level Zero headers are transitively included without pointing to the location of the headers. I managed to work around this by modifying the CMakeLists.txt file of our test to add these include directories (see the first comment of the branch).

After working around this header include issue, I encountered other issues from the UR code that maps Level Zero to UR (sycl/plugins/unified_runtime/ur/adapters/level_zero/ur_level_zero.hpp) Rather than starting to modifying this code which we don't own, and will soon disappear, instead I'm proposing that we remove these tests for now since PI is being removed (second commit of this branch). Then re-add the unit tests once there is UR unit test suite, as these tests are still in our git history to revive and port.