[SYCL][Graphs] Remove blocking wait from graph enqueue.
The current approach uses the same LO PI event for all submissions
of the same graph, and also doesn't use the wait events to
enforce dependencies on the command-list submission. By doing these
in the L0 adapter, we can remove the blocking queue wait from
our graphs submission code in the runtime.

Closes issue https://github.com/reble/llvm/issues/139