[SYCL][Graph] Expand Subgraph testing
When a user adds a subgraph we currently link all the nodes
in a subgraph to the parent graph. This causes issues in the
following cases:

* User submits sub-graph object to execute standalone, as it
  exit nodes now have sucessors to a parent graph
* Sub-graph object is added to another graph as a node (or the same
  graph again), as its original root nodes now have predecessors.

Documented these limitations in the README and created ticket
https://github.com/reble/llvm/issues/201 to resolve this.

Actions https://github.com/reble/llvm/issues/142 as this also
has a test for nested sub-graphs, which do work.