[SYCL][Doc] Expand on update() error wording for invalid device
Expand on the error wording around using `update()` when passing
a graph with a different device. This is technically already
covered on the node level by the definition of *topologically
identical*, but defining an error at the graph level as well is
more obvious to a reader.

Actions specfication feedback from
https://github.com/intel/llvm/pull/5626#discussion_r1206832653