[SYCL][Graph] Fix test-e2e host task related fails
With our change from storing a host-task in a command-group to a
shared-pointer from a unique shared-pointer we didn't cover all the
usages.

When the native kernel is called in `DispatchNativeKernel`,
the `void*` argument is currently casted to a unique_ptr
and then freed at the end of the function. However, we set this
`void*` to a raw pointer from the shared-pointer `get()`. This results in
freeing the underlying memory held by the shared-pointer in
`DispatchNativeKernel`  before the shared-pointer's reference count is
decremented to zero.

I initially tried removing the `delete` in `DispatchNativeKernel`,
which fixes `Basic/handler/run_on_host_intel.cpp` but doesn't
fix `Basic/handler/run_on_host_intel2.cpp`, because user code can result
in the command-group being deleted before the native kernel is run, causing the
shared-pointer to decrement to a zero reference count and free the memory.

The proposed fix in this patch creates a shared-pointer on the heap every time
the host-task is enqueued, this increments the reference count of the
shared-pointer for the lifetime it takes for the `DispatchNativeKernel` callback
to run, and the heap shared-pointer is then freed at the end of `DispatchNativeKernel`.

Fixes the following test-e2e tests for OpenCL CPU:
```
Basic/handler/run_on_host_intel.cpp
Basic/handler/run_on_host_intel2.cpp
DiscardEvents/discard_events_host_task.cpp
```