[SYCL][Graph] Throw exception when calling profiling
Throws an invalid exception when calling `event::get_profiling_info()` on an event returned by a graph submission.
Adds unitests checking that exceptions are throws if `event::get_profiling_info()` is called on an event returned from a queue in recording mode or an event returned by a graph submission.