[SYCL][Graph] Enable empty nodes in Subgraphs
The implementation uses the list of scheduled nodes to add a subgraph to a main graph.
However, since empty nodes are not scheduled, empty nodes were not listed in this list, resulting in inconsistent graphs when subgraph with empty node(s) were added to a main graph.
This PR fixes this issue by forcing to list empty nodes when creating the list for inserting subgraph.
It also adds unitests to check that subgraphs with empty nodes are correctly added to a main graph.