[SYCL][Graph] Enable empty nodes in Subgraphs
Changes the definition of MSchedule as it now contains all types of nodes (including empty nodes).
Empty nodes are now filtered out when creating the commandbuffer and/or enqueuing nodes to only keep their dependencies.
This PR updates a few unitests to make them compliant to the new schedule list definition.