[SYCL][Graph] Improve multithread e2e tests
Multihtread finalize.cpp e2e was written to test both finalize and submit processes.
This test was based on strong assumptions on backend's kernel ordering, and was unstable.
Since the behaviour of the finalize process in multi-threading environment is already tested by unitests,
we (re)focused this test to verify concurrent submissions.
The test is therefore renamed submit.cpp